requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.4
